requirements.txt
crispy-bootstrap5==0.7
django==4.1.5
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
djangorestframework==3.14.0
python-decouple==3.6